+++ ucab-services/backend/src/main/java/com/ucab/ucab_services/entity/Estudiante.java
@@ -2,12 +2,12 @@ package com.ucab.ucab_services.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.Column;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.PrimaryKeyJoinColumn;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Entity
 @Getter @Setter @NoArgsConstructor
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class Estudiante extends Miembro {
 
     @Column(name = "promedio", precision = 4, scale = 2)
-    private Double promedio;
+    private BigDecimal promedio;
 
     @Column(name = "escuela", length = 150)
     private String escuela;
